+++ opencascade-hs/src/OpenCascade/RWObj/Internal/Destructors.hs
@@ -1,0 +1,10 @@
+
+{-# LANGUAGE CApiFFI #-}
+module OpenCascade.RWObj.Internal.Destructors
+( deleteCafWriter
+) where
+
+import OpenCascade.RWObj.Types
+import Foreign.Ptr 
+
+foreign import capi unsafe "hs_RWObj_CafWriter.h hs_delete_RWObj_CafWriter" deleteCafWriter :: Ptr CafWriter -> IO ()