+++ opencascade-hs/src/OpenCascade/RWObj/Types.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.RWObj.Types 
+( CafWriter
+) where
+
+data CafWriter